+++ packages/Flow/src/package-test.ts
@@ -17,6 +17,7 @@ import './tests/test-ids-tests';
 import './tests/guide-tests';
 import './tests/summary-tests';
 import './tests/files-tree-tests';
+import './tests/execution-preview-tests';
 
 export let _package = new DG.Package();
 export { tests };